Feat/#27 - 채팅방 구현

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/dto/ChatResponseDto.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @Builder
 public class ChatResponseDto {
 
-    private Long chatId;
+    private Integer chatId;
 
-    private Long roomId;
+    private Integer roomId;
 
-    private Long senderId;
+    private Integer senderId;
 
     private String message;
